requirements.txt
IPython
pandas
h5py
google-generativeai
streamlit
datasets
sentence-transformers